Rename toolkit command line option
