+++ swan-pipeline/src/main/java/de/fraunhofer/iem/swan/cli/CliRunner.java
@@ -25,8 +25,8 @@ public class CliRunner implements Callable<Integer> {
     @CommandLine.Option(names = {"-f", "--feature"}, description = {"Select one or more feature sets: all, code, doc-auto or doc-manual"})
     private List<String> featureSet =  Collections.singletonList("code");
 
-    @CommandLine.Option(names = {"-l", "--learning"}, description = {"Learning modes: meka, weka, ml-plan"})
-    private String learningMode = "meka";
+    @CommandLine.Option(names = {"-t", "--toolkit"}, description = {"ML toolkit: meka, weka, ml-plan"})
+    private String toolkit = "meka";
 
     @CommandLine.Option(names = {"-s", "--srm"}, description = {"SRM: all, source, sink, sanitizer, authentication, relevant"})
     private List<String> srmClasses = Collections.singletonList("all");
@@ -61,7 +61,7 @@ public class CliRunner implements Callable<Integer> {
                 datasetJson,
                 outputDir,
                 featureSet,
-                learningMode,
+                toolkit,
                 srmClasses,
                 cweClasses,
                 exportArffData,
