+++ swan-pipeline/src/main/java/de/fraunhofer/iem/swan/cli/SwanOptions.java
@@ -15,7 +15,7 @@ public class SwanOptions {
     private String datasetJson;
     private String outputDir;
     private List<String> featureSet;
-    private String learningMode;
+    private String toolkit;
     private List<String> srmClasses;
     private List<String> cweClasses;
     private boolean exportArffData;
@@ -26,7 +26,7 @@ public class SwanOptions {
     private double predictionThreshold;
 
     public SwanOptions(String testDataDir, String trainDataDir, String datasetJson, String outputDir,
-                       List<String> featureSet, String learningMode, List<String> srmClasses,
+                       List<String> featureSet, String toolkit, List<String> srmClasses,
                        List<String> cweClasses, boolean exportArffData, boolean isDocumented,
                        int iterations, double trainTestSplit, String phase) {
         this.testDataDir = testDataDir;
@@ -34,7 +34,7 @@ public class SwanOptions {
         this.datasetJson = datasetJson;
         this.outputDir = outputDir;
         this.featureSet = featureSet;
-        this.learningMode = learningMode;
+        this.toolkit = toolkit;
         this.srmClasses = srmClasses;
         this.cweClasses = cweClasses;
         this.exportArffData = exportArffData;
@@ -45,14 +45,14 @@ public class SwanOptions {
     }
 
     public SwanOptions(String testDataDir, String trainDataDir, String datasetJson, String outputDir,
-                       List<String> featureSet, String learningMode, List<String> srmClasses, List<String> cweClasses,
+                       List<String> featureSet, String toolkit, List<String> srmClasses, List<String> cweClasses,
                        boolean exportArffData, boolean isDocumented, int iterations, double trainTestSplit) {
         this.testDataDir = testDataDir;
         this.trainDataDir = trainDataDir;
         this.datasetJson = datasetJson;
         this.outputDir = outputDir;
         this.featureSet = featureSet;
-        this.learningMode = learningMode;
+        this.toolkit = toolkit;
         this.srmClasses = srmClasses;
         this.cweClasses = cweClasses;
         this.exportArffData = exportArffData;
@@ -101,12 +101,12 @@ public class SwanOptions {
         this.featureSet = featureSet;
     }
 
-    public String getLearningMode() {
-        return learningMode;
+    public String getToolkit() {
+        return toolkit;
     }
 
-    public void setLearningMode(String learningMode) {
-        this.learningMode = learningMode;
+    public void setToolkit(String toolkit) {
+        this.toolkit = toolkit;
     }
 
     public List<String> getSrmClasses() {
@@ -185,7 +185,7 @@ public class SwanOptions {
                 ", datasetJson='" + datasetJson + '\'' +
                 ", outputDir='" + outputDir + '\'' +
                 ", featureSet='" + featureSet + '\'' +
-                ", learningMode='" + learningMode + '\'' +
+                ", learningMode='" + toolkit + '\'' +
                 ", srmClasses=" + srmClasses +
                 ", cweClasses=" + cweClasses +
                 ", exportArffData=" + exportArffData +
